fix test dataset generation
