+++ {{ cookiecutter.repo_name }}/{{ cookiecutter.module_name }}/tests/make_test_datasets.py
@@ -1,15 +1,12 @@
 from sklearn.datasets import fetch_20newsgroups
 from functools import partial
 
-from {{ cookiecutter.module_name }}.data import DataSource, Dataset, DatasetGraph, Catalog
-from {{ cookiecutter.module_name }} import workflow, paths
-from {{ cookiecutter.module_name }}.log import logger
+from src.data import DataSource, Dataset, DatasetGraph, Catalog
+from src.data.process_functions import process_20_newsgroups
+from src import paths
+from src.log import logger
 
 # Set up a 20 newsgroups dataset
-
-ds_name = '20_newsgroups'
-output_ds_name = ds_name
-dsrc = DataSource(ds_name)
 
 license = """
 Custom Academic License: "You may use this material free of charge for any educational purpose, provided attribution is given in any lectures or publications that make use of this material." As in http://kdd.ics.uci.edu/databases/20newsgroups/20newsgroups.data.html.
@@ -46,51 +43,19 @@ The current version is obtained by wrapping `sklearn.datasets.fetch_20newsgroups
 
 By default we follow the sklearn suggestion to set `remove=('headers', 'footers', 'quotes')` to avoid overfitting.
 """
+if __name__ =='__main__':
+    ds_name = '20_newsgroups'
+    output_ds_name = ds_name
+    dsrc = DataSource(ds_name)
 
-dsrc.add_metadata(contents=metadata, force=True)
-dsrc.add_metadata(contents=license, kind='LICENSE', force=True)
+    dsrc.add_metadata(contents=metadata, force=True)
+    dsrc.add_metadata(contents=license, kind='LICENSE', force=True)
 
-def process_20_newsgroups(*, extract_dir='20_newsgroups',
-                          metadata=None, unpack_dir=None,
-                          opts={"subset":"all", "remove":"('headers', 'footers', 'quotes')"}):
-    """
-    Process 20 newsgroups into (data, target, metadata) format.
+    process_function = process_20_newsgroups
+    process_kwargs = {}
 
+    dsrc.process_function = partial(process_function, **process_kwargs)
+    dsrc.update_catalog()
 
-    Parameters
-    ----------
-    unpack_dir: path
-        The interim parent directory the dataset files have been unpacked into.
-    extract_dir: str
-        Name of the directory of the unpacked files relative to the unpack_dir. Note that
-    opts: dict default {"subset":"all", "remove"="('headers', 'footers', 'quotes')"}
-        Options to pass to sklearn.datasets.fetch_20newsgroups.
-
-
-    Returns
-    -------
-    A tuple:
-        (data, target, additional_metadata)
-
-    """
-    if metadata is None:
-        metadata = {}
-
-    if unpack_dir is None:
-        unpack_dir = paths['interim_data_path']
-    else:
-        unpack_dir = pathlib.Path(unpack_dir)
-    data_dir = unpack_dir / f"{extract_dir}"
-
-    news = fetch_20newsgroups(**opts)
-
-    return news.data, news.target, metadata
-
-process_function = process_20_newsgroups
-process_kwargs = {}
-
-dsrc.process_function = partial(process_function, **process_kwargs)
-dsrc.update_catalog()
-
-dag = DatasetGraph()
-dag.add_source(output_dataset=output_ds_name, datasource_name=ds_name, overwrite_catalog=True)
+    dag = DatasetGraph()
+    dag.add_source(output_dataset=output_ds_name, datasource_name=ds_name, overwrite_catalog=True)
